directory consolidation reporting
