+++ docs/architecture/batch_data_flow.md
@@ -89,7 +89,7 @@ Mixed batches (margin + spot scenarios in the same currency group) are split int
 
 ## Report Sections: WarningsSummary
 
-`WarningsSummary` (`python/framework/batch_reporting/warnings_summary.py`) consolidates all global warnings into a single report section. Unlike other report sections, it is **always rendered** regardless of the `summary.detail` flag, but only when at least one warning is active.
+`WarningsSummary` (`python/framework/reporting/console/warnings_summary.py`) consolidates all global warnings into a single report section. Unlike other report sections, it is **always rendered** regardless of the `summary.detail` flag, but only when at least one warning is active.
 
 Current warnings:
 - **Stress test active** — lists active stress test configs grouped by signature
