+++ python/framework/reporting/console/abstract_batch_summary_section.py
