+++ python/framework/reporting/console/broker_info_renderer.py
