+++ python/framework/reporting/console/portfolio_summary.py
@@ -9,7 +9,7 @@ unified report model (#393) — `PortfolioReport` (full projection) + `PendingOr
 
 from typing import Dict, List, Optional
 
-from python.framework.batch_reporting.abstract_batch_summary_section import AbstractBatchSummarySection
+from python.framework.reporting.console.abstract_batch_summary_section import AbstractBatchSummarySection
 from python.framework.utils.console_renderer import ConsoleRenderer
 from python.framework.types.api.report_types import (
     ActiveOrderRow, AggregatedPortfolioCurrency, AggregatedPortfolioReport, AggregatedPortfolioRow,
