+++ tests/framework/reporting/test_broker_report.py
@@ -15,7 +15,7 @@ import re
 from contextlib import redirect_stdout
 
 from python.configuration.market_config_manager import MarketConfigManager
-from python.framework.batch_reporting.broker_summary import BrokerSummary
+from python.framework.reporting.console.broker_summary import BrokerSummary
 from python.framework.reporting.run_reports.broker_report_builder import (
     build_broker_report_from_batch, build_broker_report_from_session)
 from python.framework.types.api.report_types import (
