+++ tests/framework/reporting/test_profiling_report.py
@@ -14,7 +14,7 @@ import re
 from contextlib import redirect_stdout
 from datetime import datetime, timezone
 
-from python.framework.batch_reporting.profiling_summary import ProfilingSummary
+from python.framework.reporting.console.profiling_summary import ProfilingSummary
 from python.framework.reporting.run_reports.profiling_report_builder import (
     build_profiling_report_from_batch)
 from python.framework.reporting.run_reports.report_aggregators import aggregate_profiling
